+++ RandomMoonFX/Plugin.cs
@@ -1,7 +1,8 @@
 ﻿using BepInEx;
+using BepInEx.Bootstrap;
 using HarmonyLib;
 using System.Collections.Generic;
-using System.Linq;
+
 
 namespace RandomMoonFX
 {
@@ -10,7 +11,7 @@ namespace RandomMoonFX
     {
         const string GUID = "zigzag.randommoonfx";
         const string NAME = "RandomMoonFX";
-        const string VERSION = "1.2.5";
+        const string VERSION = "1.3.0";
 
         public static Plugin instance;
         private readonly Harmony harmony = new Harmony(GUID);
@@ -19,24 +20,36 @@ namespace RandomMoonFX
         private readonly int GordionID = 3;
         public float AnimationTime = 1.5f;
         public bool IsStarting = false;
+        public Terminal? terminal;
+        public int terminalCostOfItems = -5;
         public List<string> VisitedMoons = new List<string>();
 
+        [System.Diagnostics.CodeAnalysis.SuppressMessage("CodeQuality", "IDE0051")]
         void Awake()
         {
             instance = this;
             SetParameters();
-            harmony.PatchAll();
+            if (config.ActivateRandomMoons.Value)
+            {
+                harmony.CreateClassProcessor(typeof(StartOfRoundPatch), true).Patch();
+                harmony.CreateClassProcessor(typeof(StartMatchLeverPatch), true).Patch();
+            }
+            if (config.ActivateFreeMoons.Value)
+            {
+                harmony.CreateClassProcessor(typeof(TerminalPatch), true).Patch();
+            }
             Logger.LogInfo("RandomMoonFX is loaded !");
         }
 
         private void SetParameters()
         {
             config = new Config(base.Config);
-            var assemblies = System.AppDomain.CurrentDomain.GetAssemblies();
-            if (config.CelestialTintAnimation.Value && assemblies.Any(assembly => assembly.FullName.StartsWith("CelestialTint")))
-            {
+            if (config.ChameleonAnimation.Value && Chainloader.PluginInfos.ContainsKey("butterystancakes.lethalcompany.chameleon"))
+                AnimationTime = 7.5f;
+            if (config.CelestialTintAnimation.Value && Chainloader.PluginInfos.ContainsKey("CelestialTint"))
                 AnimationTime = 4f;
-            }
+            if (config.AnimationTimeOverride.Value >= 0f)
+                AnimationTime = config.AnimationTimeOverride.Value;
         }
 
         public bool LastDayOfQuota()